fix(chat): add useIsPublicationReview hook
